Add bottle unneeded to tap formulae (#15)
These formula will not be bottled because we unpack the built artifacts
directly. Additionally, this escapes some detection on Catalina where it
appears the detection on whether or not the formula needs to be
installed from source is busted from failing to detect that the
development tools are installed. By marking these formula correct as
bottle unneeded, we circumvent that problem.